Add files via upload
This document is the Final Bachelor Thesis (TFG) report, which presents the research, methodology, and results obtained in the characterization of materials in the millimeter-wave frequency band. It provides a detailed explanation of the techniques used to extract key electromagnetic properties, such as complex refractive index, relative permittivity, and permeability.